Adding links to newlogin and newsignup. Login and Signup forms now work
