+++ app/static/css_styles/common_styles.css
@@ -1,3 +1,10 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  /* Unified Box Model */
+}
+
 body {
     /* Page background color and font settings */
     background: #f4f4f4;
@@ -70,7 +77,7 @@ body {
     font-size: 50px;
     display: inline-block;
     border-bottom: 10px solid blue;
-    padding-bottom: 10px;
+    padding-bottom: 0px;
 }
 
 .subtitle{
@@ -102,27 +109,21 @@ body {
 .title_container{
     padding-left: 30px;
     padding-right: 30px;
-    background-color: white;
+    background-color: #f4f4f4;
 }
 
 .login_container{
     background-color: darkgreen;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 30px;
+    padding: 30px;
 }
 
 .signup_container{
     background-color: darkblue;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 30px;
+    padding: 30px;
 }
 
 .about_container{
     background-color: #e8d315;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 30px;
+    padding: 30px;
     height: 100%;
 }
