+++ app/static/css_styles/loginButtonStyle.css
@@ -18,7 +18,7 @@ body {
   /* Absolute positioning of child elements */
   width: 100px;
   /* Container width, set according to button size */
-  margin: 100px auto;
+  margin: 50px auto;
   /* center the container */
   text-align: center;
   /* Internal text is centered horizontally */
