+++ app/static/css_styles/upload_styles.css
@@ -1,8 +1,6 @@
 .upload_container {
     background-color: darkgreen;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 30px;
+    padding: 30px;
     position: relative;
 }
