+++ app/static/css_styles/visualise_styles.css
@@ -1,20 +1,14 @@
 .movie_container {
     background-color: darkgreen;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 30px;
+    padding: 30px;
 }
 
-.movie_image_container{
+.movie_image_container {
     background-color: darkblue;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 30px;
+    padding: 30px;
 }
 
-.grey_container{
+.grey_container {
     background-color: grey;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 30px;
+    padding: 30px;
 }
