+++ app/templates/base.html
@@ -10,8 +10,12 @@
     {% if stylesheet_file is defined %}
     <link rel="stylesheet" href="{{ url_for('static', filename=stylesheet_file) }}">
     {% endif %}
+    <link rel="stylesheet" href="{{ url_for('static', filename='css_styles/loginButtonStyle.css') }}">
 </head>
 <body>
+
+
+  
     <div class="website_title"><h1>Movie Tracker</h1></div>
 
     <nav class="navbar">
@@ -21,9 +25,11 @@
             <li {% if underlined_tab_index == 3%}class="underlined_tab"{% endif %}><a href="/visualiseData">Visualise Data</a></li>
             <li {% if underlined_tab_index == 4%}class="underlined_tab"{% endif %}><a href="/shareData">Share Data</a></li>
             {% if current_user.is_authenticated %}
-                <li {% if underlined_tab_index == 5%}class="underlined_tab"{% endif %}><a href="/logout">Logout ({{ current_user.username }})</a></li>
+            <li {% if underlined_tab_index == 5%}class="underlined_tab"{% endif %}><a href="/logout">Logout ({{ current_user.username }})</a></li>
             {% else %}
+            <!--
                 <li {% if underlined_tab_index == 6%}class="underlined_tab"{% endif %}><a href="/signup">Sign Up</a></li>
+                -->
             {% endif %}
         </ul>
     </nav>
