+++ app/templates/login_modal.html
@@ -15,25 +15,23 @@
         <button type="button" class="closeButton"onclick="window.location.href='/'">Close</button>
         <!-- Centered container for pop-up content -->
         <div class="center-login-page">
-            <form action="" class="center-login-page" method="post">
+            <form action="{{ url_for('login') }}" class="center-login-page" method="POST">
                 <!-- Input area: Username and Password -->
                 <div class="tab-form">
                     <!-- Username input item -->
-                    <div class="form-item">
+                    <div action="{{ url_for('login') }}" method = "POST" class="form-item">
                         <div class="form-info">Username</div>
-                        <input type="text" id="username" name="username" placeholder="Please enter username" required>
+                        <input type="text" name="username" placeholder="Please enter username" id="login_username" required>
                     </div>
                     <!-- Password input item, including the "Forgot Password" button -->
                     <div class="form-item">
                         <div class="form-info">Password</div>
-                        <input type="password" name="password" id="password" placeholder="Please enter password"
-                            required>
+                        <input type="password" name="password" placeholder="Please enter password" id="login_password"required>
                         <button type="button"onclick="window.location.href='{{ url_for('forgetPassword') }}'">Forget password?</button>
                     </div>
                 </div>
                 <!-- Submit and Register Buttons -->
-                <button type="submit">Log In</button><button type="button"onclick="window.location.href='{{ url_for('newsignup') }}'">Sign Up</button>
-
+                <button type="submit" value="login">Log In</button><button type="button"onclick="window.location.href='{{ url_for('newsignup') }}'">Sign Up</button>
             </form>
         </div>
