+++ app/templates/shareData.html
@@ -4,6 +4,8 @@
 
 {% set stylesheet_file = 'css_styles/share_styles.css' %}
 
+{% include 'loginButton.html'%} <!-- This is the newly designed login button-->
+
 {% block content %}
 
 <!--Share Data section-->
