+++ app/templates/uploadData.html
@@ -4,6 +4,8 @@
 
 {% set stylesheet_file = 'css_styles/upload_styles.css' %}
 
+{% include 'loginButton.html'%} <!-- This is the newly designed login button-->
+
 {% block content %}
 <div class="body_container">
     <div class="centered_container">
