+++ app/templates/visualiseData.html
@@ -4,6 +4,8 @@
 
 {% set stylesheet_file = 'css_styles/visualise_styles.css' %}
 
+{% include 'loginButton.html'%} <!-- This is the newly designed login button-->
+
 {% block content %}
 <!--Visualise Data section-->
 <div class="body_container">
@@ -13,26 +15,27 @@
                 <h1 class="title">Visualise Data</h1>
             </div>
         </div>
-        <div class="centered_div row_direction">
-            <div class="movie_container">
-                <h2 class="subtitle white_font">Movie 1</h2>
+            <div class="centered_div row_direction movie">
+                <div class="movie_container">
+                    <h2 class="subtitle white_font">Movie 1</h2>
+                </div>
+                <div class="movie_image_container">
+                    <h2 class="subtitle white_font">Movie Image</h2>
+                </div>
+                <div class="grey_container">
+                    <h2 class="subtitle white_font">Movie Details</h2>
+                </div>
             </div>
-            <div class="movie_image_container">
-                <h2 class="subtitle white_font">Movie Image</h2>
-            </div>
-            <div class="grey_container">
-                <h2 class="subtitle white_font">Movie Details</h2>
-            </div>
-        </div>
-        <div class="centered_div row_direction">
-            <div class="movie_container">
-                <h2 class="subtitle white_font">Movie 2</h2>
-            </div>
-            <div class="movie_image_container">
-                <h2 class="subtitle white_font">Movie Image</h2>
-            </div>
-            <div class="grey_container">
-                <h2 class="subtitle white_font">Movie Details</h2>
+            <div class="centered_div row_direction">
+                <div class="movie_container">
+                    <h2 class="subtitle white_font">Movie 2</h2>
+                </div>
+                <div class="movie_image_container">
+                    <h2 class="subtitle white_font">Movie Image</h2>
+                </div>
+                <div class="grey_container">
+                    <h2 class="subtitle white_font">Movie Details</h2>
+                </div>
             </div>
         </div>
     </div>
